lensden/requirement.txt
djangorestframework==3.9.4
Markdown==3.1.1
pkg-resources==0.0.0
absl-py==0.8.1
apturl==0.5.2
asgiref==3.2.5
asn1crypto==0.24.0
astor==0.8.1
Brlapi==0.6.6
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0.4
django-environ==0.4.5
django-filter==2.2.0
django-json==0.27
djangorestframework==3.11.0
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
httplib2==0.9.2
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
jsonparser==1.0
kazam==1.4.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mtcnn==0.1.0
netifaces==0.10.4
numpy==1.17.4
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.2.30
pexpect==4.2.1
Pillow==6.2.1
protobuf==3.11.1
psycopg2==2.8.4
pycairo==1.16.2
pycocotools==2.0.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.5
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==5.2
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
scipy==1.3.3
SecretStorage==2.3.1
simplejson==3.13.2
six==1.13.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
